FFixedd

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	char d;
 	float f;
 
-	printf("Size of char: %zu byte(s)\n", sizeof(d));
-	printf("Size of an int: %zu byte(s)\n", sizeof(a));
-	printf("Size of a long int: %zu byte(s)\n", sizeof(b));
-	printf("Size of long long int: %zu byte(s)\n", sizeof(c));
-	printf("Size of a float: %zu byte(s)\n", sizeof(f));
+	printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(d));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(a));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
